+++ tests/e2e/privateProfile.spec.ts
@@ -1,12 +1,11 @@
+import Lecturer from "@/api/api_models/user_management/Lecturer";
 import Student from "@/api/api_models/user_management/Student";
 import { Account } from "@/entities/Account";
-import Lecturer from "@/api/api_models/user_management/Lecturer";
 import { Country } from "@/entities/Country";
 import { getRandomMatriculationId, deleteUsers, createUsers, getRandomizedGovernmentId } from "./helpers/UserHelper";
 import { getMachineUserAuth, loginAsUser, logout } from "./helpers/AuthHelper";
 import { navigateToPrivateProfile } from "./helpers/NavigationHelper";
 import { UserWithAuth } from "./helpers/UserWithAuth";
-import { Role } from "@/entities/Role";
 
 describe("Change Profile Information", () => {
     const random = Math.floor(Math.random() * 9999);